// src/office-viewer/services/presentation/types.ts
export type TextStyle = {
  fontFamily?: string;
  fontSize?: number;
  bold?: boolean;
  italic?: boolean;
  underline?: boolean;
  strike?: 'none' | 'sngStrike' | 'dblStrike';
  smallCaps?: boolean;
  allCaps?: boolean;
  color?: string;
  opacity?: number;
  align?: 'left' | 'center' | 'right' | 'justify';
  verticalAlign?: 'top' | 'middle' | 'bottom';
  writingMode?: 'horizontal-tb' | 'vertical-rl' | 'vertical-lr';
  fit?: 'none' | 'shrinkText' | 'resizeShape';
  marginLeft?: number;
  marginRight?: number;
  marginTop?: number;
  marginBottom?: number;
  lineHeight?: number;
  spaceBefore?: number;
  spaceAfter?: number;
  textIndent?: number;
  charSpace?: number;
  baseline?: number;
  bullet?: TextBulletStyle;
};

export type TextBulletStyle = {
  char?: string;
  color?: string;
  size?: number;
  none?: boolean;
};

export type TextRun = {
  text: string;
  style?: TextStyle;
};

export type TextParagraph = {
  runs: TextRun[];
  style?: TextStyle;
  level?: number;
  bullet?: TextBulletStyle;
};